Restructured code for ChangeSet
